requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
openai>=1.35.0
python-dotenv==1.0.0
requests==2.31.0